+++ wcomponents-core/src/main/java/com/github/bordertech/wcomponents/WMessageBox.java
@@ -9,6 +9,7 @@ import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 
 /**
  * <p>
@@ -147,7 +148,9 @@ public class WMessageBox extends AbstractWComponent implements AjaxTarget, Subor
 	 * @param type the messageBox type, one of {@link #SUCCESS}, {@link #INFO}, {@link #WARN} or {@link #ERROR}.
 	 */
 	public void setType(final Type type) {
-		getOrCreateComponentModel().type = type;
+		if (type != getType()) {
+			getOrCreateComponentModel().type = type;
+		}
 	}
 
 	/**
@@ -164,8 +167,13 @@ public class WMessageBox extends AbstractWComponent implements AjaxTarget, Subor
 	 * @param args optional arguments for the message format string.
 	 */
 	public void setTitleText(final String title, final Serializable... args) {
-		MessageModel model = getOrCreateComponentModel();
-		model.title = I18nUtilities.asMessage(title, args);
+		Serializable currTitle = getComponentModel().title;
+		Serializable titleToBeSet = I18nUtilities.asMessage(title, args);
+
+		if (!Objects.equals(titleToBeSet, currTitle)) {
+			MessageModel model = getOrCreateComponentModel();
+			model.title = titleToBeSet;
+		}
 	}
 
 	/**