+++ wcomponents-core/src/test/java/com/github/bordertech/wcomponents/WLabel_Test.java
@@ -3,6 +3,8 @@ package com.github.bordertech.wcomponents;
 import junit.framework.Assert;
 import org.junit.Test;
 
+import java.io.Serializable;
+
 /**
  * WLabel_Test - unit tests for {@link WLabel}.
  *
@@ -101,5 +103,13 @@ public class WLabel_Test extends AbstractWComponentTestCase {
 		label.setSanitizeOnOutput(true);
 		label.setEncodeText(false);
 		Assert.assertEquals("Expect output to be sanitized", "content", label.getText());
+	}
+
+	@Test
+	public void testDuplicateComponentModels() {
+		WLabel wLabel = new WLabel("tesLabel");
+		assertNoDuplicateComponentModels(wLabel, "text", "testText", new Serializable[]{});
+		assertNoDuplicateComponentModels(wLabel, "hint", "testHint", new Serializable[]{});
+		assertNoDuplicateComponentModels(wLabel,"sanitizeOnOutput", true);
 	}
 }