+++ wcomponents-core/src/test/java/com/github/bordertech/wcomponents/WMultiFileWidget_Test.java
@@ -3,11 +3,11 @@ package com.github.bordertech.wcomponents;
 import com.github.bordertech.wcomponents.WMultiFileWidget.FileWidgetUpload;
 import com.github.bordertech.wcomponents.file.FileItemWrap;
 import com.github.bordertech.wcomponents.util.mock.MockFileItem;
+import junit.framework.Assert;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.fileupload.FileItem;
-import org.junit.Assert;
 import org.junit.Test;
-
+import java.awt.Dimension;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -307,6 +307,19 @@ public class WMultiFileWidget_Test extends AbstractWComponentTestCase {
 	public void testMaxFileSizeAccessors() {
 		assertAccessorsCorrect(new WMultiFileWidget(), "maxFileSize", (long) 10240000, (long) 1,
 				(long) 2);
+	}
+
+	@Test
+	public void testDuplicateComponentModels() {
+		WMultiFileWidget multiFileWidget = new WMultiFileWidget();
+		assertNoDuplicateComponentModels(multiFileWidget,"maxFileSize", 2012312312);
+		assertNoDuplicateComponentModels(multiFileWidget, "maxFiles", 123);
+		assertNoDuplicateComponentModels(multiFileWidget, "newUpload", true);
+		assertNoDuplicateComponentModels(multiFileWidget, "useThumbnails", true);
+		assertNoDuplicateComponentModels(multiFileWidget, "thumbnailPosition", WLink.ImagePosition.SOUTH);
+		assertNoDuplicateComponentModels(multiFileWidget, "thumbnailSize", new Dimension(22,33));
+		// TODO: See issue #1574 https://github.com/BorderTech/wcomponents/issues/1574
+//		assertNoDuplicateComponentModels(multiFileWidget, "fileUploadRequestId", "testId"); // No such method exception as it's a private method
 	}
 
 	@Test