+++ wcomponents-core/src/test/java/com/github/bordertech/wcomponents/WWindow_Test.java
@@ -2,12 +2,13 @@ package com.github.bordertech.wcomponents;
 
 import com.github.bordertech.wcomponents.servlet.WebXmlRenderContext;
 import com.github.bordertech.wcomponents.util.mock.MockRequest;
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.io.StringWriter;
 import junit.framework.Assert;
 import org.junit.Test;
 import org.xml.sax.SAXException;
+
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.io.StringWriter;
 
 /**
  * Window_Test - unit tests for {@link WWindow}.
@@ -122,5 +123,25 @@ public class WWindow_Test extends AbstractWComponentTestCase {
 		window.handleRequest(request);
 		window.paint(new WebXmlRenderContext(new PrintWriter(writer)));
 		Assert.assertEquals("Content should have been painted", 1, content.getPaintCount());
+	}
+
+	@Test
+	public void testDuplicateComponentModels() {
+		WWindow window = new WWindow();
+		assertNoDuplicateComponentModels(window, "title", "testTitle");
+		assertNoDuplicateComponentModels(window, "height", 223);
+		assertNoDuplicateComponentModels(window, "width", 223);
+		assertNoDuplicateComponentModels(window, "top",2);
+		assertNoDuplicateComponentModels(window, "left",3);
+		assertNoDuplicateComponentModels(window, "showMenuBar", true);
+		assertNoDuplicateComponentModels(window, "showToolbar", true);
+		assertNoDuplicateComponentModels(window, "showLocation", true);
+		assertNoDuplicateComponentModels(window, "showStatus", true);
+		assertNoDuplicateComponentModels(window, "resizable", false);
+		assertNoDuplicateComponentModels(window, "scrollable", true);
+		// TODO: See issue #1574 https://github.com/BorderTech/wcomponents/issues/1574
+//		assertNoDuplicateComponentModels(window, "state", 3); // No such method exception - protected modifier
+//		assertNoDuplicateComponentModels(window, "targeted", true); // No such method exception - protected modifier
+		assertNoDuplicateComponentModels(window, "step", 5);
 	}
 }